[TASK] Create an itemtype factory
The factory should be used with a similar construction to FAL's
retrieve-or-create instantiation method.

By simply requesting an itemtype object via itemtype string, the factory
needs to be able to:
- check its local cache, or
- use the repository to find it, or
- instantiate an object, use the repository to store it, and put it in
local cache with its corresponding uid set

While creating this for the particular valueobject that is itemType, we
know this construction will be needed later on for Fields as well. With
that in mind, I need to create a FactoryAbstract as well.

Change-Id: I15655ee6bdb5f3b43d88d639c8a795b4097a2f20