[TASK] Throw staticUidInsertion exception in Repository->insertRecord
When an 'uid' property is given to the insert-data, it is currently
being unset. This way, it won't produce serious issues in the database,
however, one may assume his action succeeded with the initially provided
uid as a result.

Let's face it, this is not a transparent way of handling the situation,
nor does it immediately discourage said practice. Hence we have to throw
an exception when an uid is detected with the insertion data, showing
this is not accepted.

Change-Id: Ibd383b27f00f8de4dad759969f7def3f0fc187c6