[TASK] Create Extbase Storage Handler for importer
Handles storage of parsed import via Extbase Domain Models,
Repositories, Factories and the PersistenceManager. Note that while this
is the correct way to implement storage handling, it is just not
realistic to use performance-wise, in combination with the
OneFileStreamingParser (default).

However, its development contributes to the NEOS version of the
extension considerably.

Change-Id: Icbf08883d32db8e553c030b81ad3de395fb3be49