[TASK] Create Classic Storage Handler for importer
Handles storage of parsed import via direct queries. Although this
implementation has an enormous speed-benefit as opposed to other
StorageHandlers, it will not update TYPO3-specific reference fields,
with the exception of FileReferences. However, the extension doesn't
really rely on them to function properly.

This storage handler improves performance by a factor of 3.1 compared to
the extbase storage handler, which can make all the difference for large
import files.

Change-Id: Ie9419d77551d91d78af2cc46e5b06e5fbf11cc34