[TASK] Create Database-specific Query Providers
We've already introduced the QueryProviderInterface with the Classic
Storage Handler for the importer. Let's add the required methods.

First up is the upsertQuery, which was heavily utilized by the old
decospublisher, solely through the MySQL implementation. However,
because the importer can't use it the same way as the old one did for
various tables, because we can't provide unique constraints everywhere
without considering the impact this has on TCA and pid and deleted. So
we also need a fallback, like a CompatiblityQueryProvider that uses
DatabaseConnection provided methods only to achieve our goal.

Meanwhile, this interface and the implementations can serve as the host
of any database-specific query, so that we may more easily ready the
extension for other database-engines when the need arises.

Change-Id: I392c062c1ddc64a5997d0a40c0e48e0f4db65f4b