[TASK] Set Auto-updating imports through task
In the old version, you had to set auto-updating as a property of the
XML record. Not very flexible. So instead, we're making a selection of
imports (regardless of pid) in the importer task.

This way, one has the added flexibility of maintaining different
update-frequencies for different imports, and also no longer requires
you to edit the import record for setting a property that only has
context in a single task, as it is now done from within the task itself.

Naturally, this allows us to completely remove the auto_update property.

Change-Id: I84f15af16f1357742e9e5a664d7ac8aed25eb34e