[TASK] Expand Migrate DatabaseService with migrated_uid support
All our migration methods rely on removing old data, but the move to a
new extension comes with the motivation to keep the old data until
you've got the new extension up and running so that it can replace AFTER
you're done migrating and configuring.

This means we have to create migration methods for database that rely on
setting a migrated_uid property instead. This property will be
introduced into the classic decospublisher's final version.

At the same time, we have to support the ability to use the fileService
to set FileReferences. Take note that this is a performance killer if
using the official api's through FileReferenceRepository. Regardless, it
is the easiest way to get things migrated correctly.

Also, we need more efficient ways of retrieving reference uid's to set
for e.g. newly introduced valueObjects (item_type, field) and migrated
records (migrated_uid).

Change-Id: I425d3f14b9560b1cd3f64c6d8a19fed73051ba70