[TASK] Create a basic working plugin
The plugin needs to be able to display the most simple document stream
currently used by the main sponsor with decospublisher.

There should be a query builder service covering all the basics, as well
as the ability to assign options for query and for display.

For now, the configuration can be set via an internal array, as the
actual configuration processing will be a task in itself.

Change-Id: Ida482f01c2f0f6c0967cafdbd352eb19ed4ccd00