[TASK] Refactor QueryBuilder configuration
Query Configurations were completely array based, but this gave us some
complications in having to check for consistency everywhere, and made it
necessary to ease configuring by creating "provider" methods that would
do most of the formatting of the arrays for you.

While the consistency issue and lack of ease came up, there were other
problems that I parked for later:
- Parameters are very fickle, and there was to be taken absolute care of
providing the right parameter into the right array element, to match the
order in $queryConfiguration. This is an absolute mess and can't be
relied upon on any way better than the original tx_decospublisher
clumsily provided its query alterations. (strpos/str_replace anyone?)
- There was no way to consistently build the option interface for each
leveltype, when I also needed to be able to detect aliases for
alterations. I chose to park this for much later, not giving much
thought about it.
- Constraints and separation of field and table aliases is not
consistent between select, from and where objects. Makes it all the more
prone to errors.
- The new PaginateService (WIP) needs a way to execute parameterized
queries to retrieve the amount of rows. After all, we can't resolve the
item rows to actual table rows as they're flexible constructs, leaving
the extbase/flow QueryResultInterface method useless to our needs. The
problem: it can't without working with a PreparedStatement object, which
means my entire configuration api was fucked unless I could provide both
statement and parameters in an elegant class I could feed to my
repository while still maintaining configurability.

A solution to some of these problems: named parameters. But that would
make the array construct of the parameters completely useless, which in
turn meant we had to redo some parts of it.

Then it occurred to me: why not redo every bit of it as a
class/method-based api. It would solve ALL these problems now and for in
the future, providing a solid basis we can rely on.

This task sets out to replace the ENTIRE configuration API with classes
to serve the same purpose, albeit more flexible, less error prone and
less work to validate.

Change-Id: I31109070bd5b2050edec5d14ac41759c4f540f5e