[TASK] Create a PaginateService that applies query LIMIT en OFFSET
The PaginateService should work with test-code to provide the
query-specific alterations needed to limit the resulting items. The
actual pagebrowser is a different task.

Since there are definite performance issues in calculating pagination
because of having to prepare a Statement object multiple times, we have
think of smart ways to cache the result. However, that is not part of
this task, so for this task, I have decided I can make a temporary api
to apply said LIMIT and OFFSET values to a query object, as long as I
mark them as such.

Therefore: note that the way PaginateService applies the LIMIT and
OFFSET values, is not necessarily going to be final. The aim of this
task is just to provide functional code.

Change-Id: I0fa579bac1378d4a5b8c2b7792cc5febf2b0917f