[TASK] Create a PageBrowser VH
It should utilize the PaginateService, which was configured by the
QueryBuilder, to determine the exact PageBrowser settings.

If possible, it should have its own template (partial?) to isolate it as
much as possible from the actual extension templates. This should also
make it easier to copy it to other extensions if the need ever arises
(though I should utilize the fluid paginatewidget if I can work with
QueryResults instead of arrays)

Some settings should be set from template, like placement, or adjusting
the scaling parameters. Although pretty much everything should have a
default value.

Change-Id: Ibc7617ae158b5eb0dd38939754900b6e7b3404c2