[TASK] Optimize pagebrowser
The pagebrowser renders at least one template section multiple times
with the same input arguments. We already know that the render VH does
not cache anything, so there is something to be gained.

Also need to determine if using the format.raw VH around {content} in
the template isn't too wasteful on performance, as opposed to i.e. using
an old-fashioned marker.

Change-Id: Ia83b5e28e686b3f8cbeffdea92bc899c2d5a0f92