[TASK] Catch missing filepath errors on import
These errors only affect a single itemblob, so why should it stop the
entire import process? It shouldn't! So catch it, register the error,
and continue on with the next.

Afterwards, display/log the error with any other error.

Also, Importer's ClassicStorageHandler does not check for empty
item_keys.. Should that one stop the import process?

Change-Id: I9f16a58699de1a7cad245cd281f5b3f2bbf72f9d