[TASK] Improve Query api with get-or-create methods
Query object has a createContent method, Content object has createField
method, Field object has createFrom method. What happens when you want
to get one of those by their id? We only have the create* methods, not a
specific get*(id) method. So yeah, that's a problem for the proposed
flexibility our current api offers.

Instead, why not create a get*(id) method for all of these, that returns
the object if it exists, but otherwise creates it first on the given id?

Change-Id: I2fe2f1f23388b3203efe9dcee23afa6d7787564b