[TASK] QueryConfigurator needs to resolve constraint values/operators
When we moved away from the array approach of Query Configuration, we
temporarily disabled all constraint checks that were relevant voor
FilterItems.

In our current example views, this still works, because all FilterItems'
constraints applied are special values that don't require
parameterizing. In the BIS example view we're currently trying to
create, we will have FilterItems' constraints that do need it.

Probably the easiest way is to just force validation of the value in the
QueryConfigurator's transformConstraint() method, so that even when a
value was not added to a parameter, it is forced to be one here.

Change-Id: I08410c66b8f9a6ad8baf33fe1f8ec865d0547e29