[TASK] Tackle the FilterView problem
In tx_decospublisher, we had the filterView option to determine that a
level is to be filtered by another level's argument.

Ideally, we want to provide a much easier way to configure this: link a
level to a another lever, and it is filtered automatically. However,
once we have a TCA configuration, we can perform all sorts of processing
on options chosen, so if filterView is still done through options, we
shouldn't have to concern ourselves about this, because we could e.g.
transform a checkbox selection to the appropriate options.

We know that a filterView is essentially no different from any other
filter. It has a default operator, an itemfield to compare itself to,
and a value, which in this case is taken from a GET parameter.

So why not just expand the FilterItems QueryOption to support a
'parameter' element in a filter? This would allow us to utilize what is
already there in functionality, and have all of its flexibility open up
to the FilterView problem.

Change-Id: Ia483b510d1643227011a70573ad44f1dda421518