[TASK] Add wrapping capabilities for constraints
These can be uniform to SELECT's $wrap, so we might benefit from a
WrapContainer abstract.

In addition, a container also needs a wrap for just the local field, to
support every use case required for our yearly pagination.

QueryConfigurator could then have a transformWrap() method that can be
applied to all three cases.

Change-Id: Id3eb567109b470af5b430a57913369518c20e512