[TASK] Allow overriding QueryPart wrap divider
The default wrapdivider is hardcoded as '|'. But yearly pagination is to
use regular expressions within parts that will receive wrap, and those
can (and will) contain '|' as an OR operator.

The divider needs to be settable (and gettable) while maintaining the
default.

A constraint can use the same wrapDivider for both normal wrap and
wrapLocal.

Also, options already using wrap, will need to replace the divider in
their wrap with a getWrapDivider() call. (I'm looking at you,
DateConversion!)

Change-Id: I01ade52902a81ab805f462d7599f05b26a89c592