[TASK] Allow pagination to NOT set resultCount without issues in VH
Our old yearly pagebrowser logic did not set a result count, as it would
require an additional statement creation and execution. I could do it of
course, but if it isn't necessary, it is a waste of resources and
performance. I'm inclined to not add it, at least for now.

The Pagebrowser VH will from now on detect if there is no resultCount,
and if so, disables its template configuration.

Change-Id: I69be01b0dd8387876dd4d178106210b06821c13a