[TASK] Add support for yearly-pagination
Yearly pagination reveals to us that the PaginationService can't be
Query object agnostic: we need the Query Object IN the service, so
providing just the statement no longer cuts it.

The Yearly pagination will be based on the original implementation from
tx_decospublisher, only it will be more robust by utilizing our Query
object. It will also be more costly in performance than it used to be.
But this comes with much greater maintainability, as was the aim of our
Query object.

The Pagebrowser VH also needs to support separate page labels. This
requires us to refactor the pagination configuration of the VH, so that
it utilizes not just a number, but an array with a number and label
element per page. The mapping of these labels is to be provided by the
PaginationService. In case of the default pagination, it can be left
empty, but in case of the yearly, we simply need it to contain the
resulting array of the year-clone of the Query object, which contains a
mapping to pageNumber => year. This translates perfectly to our desired
pageNumber => label.

Change-Id: Iea57b27ed0217e757be78ff72188fb681e452676