[TASK] Create RestrictByParentId option
It should give us the ability to create what was previously known as a
childView.

It seems cumbersome to have to do this from an option, while it is such
a common use-case, but I've also decided that we're going to change the
way common keys like field, blob and order are going to be handled.
Shortcuts will be accepted in user-configuration, but will all be
translated to options with some default arguments.

Change-Id: I00c965636df3fa6e4b6eb054ca581cc3545c42d9