[TASK] Basic CLI migration support
The real task is adding out-of-the-box CLI support to the ExtUpdate
library, which should repeat the tasks from the normal ExtUpdate routine
until either finished or an error occurred.

This support can be very basic, it just needs to work.

After we've established it can work and it suffices, we will refactor.

Change-Id: Ibde90be0a17fb2285ac30b7a346f52ce3e0a80bf