[TASK] Support multitable joins
We need multi table joins for the parent-in-parent query option. Why?
Because its constraints need to be paired to prevent the relation tables
from becoming a copy of the RestricyBy => it relation.

Easiest way to do it, is to change From's table mechanism to a tables
array with alias => table pairing. Throw in some methods like
addTable($table, $alias) and allow initialization with normal arrays (to
speed things up), and we have all the required flexibility with minimal
necessary changes in our current code-state.

Change-Id: Ia0fa18fbd1c5aa61a68411fcba48651f409db435