[TASK] Streamline support for QueryOptionService to retrieve index
RenderOptionService has it, so why not give it to QueryOptionService
too? ParentInParent QueryOption uses a hacky way to determine the index,
yet it's just the first of many that need it. So there are use-cases to
that could benefit by streamlining the retrieval of the index.

Doing so will require also some more refactoring on the passing of
$helperVar in the OptionServiceAbstract, because it doesn't let us give
more than a single helperVar, while this change would effectively
require us to have 2 in Query context.

The cleanest way to achieve this, is by passing the QueryOptionService
along with the options, just as is being done with RenderOptionService.
The OptionServiceAbstract can easily contain anything related to
$this->index, as both overriding classes need it.

We can get rid of "helperVar" entirely, by just passing $this from
within the abstract's executionOption() method.

Change-Id: I846094322afa05f8428e30684ac765f389f887e9