[TASK] Add CustomImage RenderOption
The option should at least support providing a relative path to the
image we want to see. It should also allow to differentiate between
replacing an empty field or one with content or relation.

Since we don't need to be exclusive about the size of the image, we can
use it as a general option for both icons and normal images alike, while
still replacing the old decospublisher icon option.

Change-Id: I91f7ad91c2b693adfcf8240661a0748ee537be5a