[TASK] Add FilterRelations QueryOption
Should work in similar fashion to FilterItems, so they could share an
abstract, also considering there might be more Filter* options in the
future.

The option is to be applied at least on column level, which we can base
for the larger part on FilterItems 's row implementation.

Also note that for options like these to work, we need options like
ParentInParent to not name their field and from entries, so these can
easily be retrieved and altered without having to know some obscure
naming scheme.

Change-Id: I971a1477d6b0c330f766b6df24f48fea4836a478