[TASK] LinkLevel should be able to link to relations
I imagine like we have linkItem now, we can use something like
linkRelation.

Since this can produce situations where there is nothing to link, we
need a failsafe that returns if there is no linkValue.

Also, since we can link on something else than content now, for fields
that don't necessarily return a content value, we need to check for an
empty content and substitute those with some default value.

Change-Id: Ic263e387ea53be696cb157e3edda68d8bfa15a45