[TASK] Rewrite the Render VH to be able to build and alter TAGS
We want to be able to add tag attributes to various tags. But if there
are multiple RenderOptions that produce an anchor tag, does that mean
that every one of those options should have an additional argument for
setting the title? The answer is NO. Consider that we'de have to do that
for EVERYTHING. In Decospublisher this was already a mess.

So an option to add tag attributes seems best. But wait, our options
alter a content string. Does the new option have to do all sorts of preg
matching or strposing to replace, append, or whatever the fuck it needs
to do with attributes? DEAR GOD NO. Decospublisher did that for query
adjustments and it was a COMPLETE MESS. Things would break easily if you
didn't know what you were doing. NO. FOR THE LOVE OF GOD, NO. Just look
at how much better it is now.

So.. we want Tag objects? Which can be adjusted to set or remove
attributes in? YES! These can be inspired by the TagBuilder object from
Fluid. Except I want a TagBuilder class to generate a Tag object.

Do that, and we open up a way to set hiding of fields through options!
Yes, let's add that too, so that function of decospublisher is covered
as well.

Change-Id: Ic07b7e3db66c5d622752e8a08d4527b19dbd35c3