[TASK] Add a fallback for old exports' lack of BLOB SEQUENCE
Since old exports didn't have a SEQUENCE field for BLOBs, we cannot use
it to determine the latest version of the document's files.

However, these cases usually do maintain a DOCUMENT_DATE field for their
BLOBs, which often has a relevant date/time value. We could use that as
a fallback to determine the order of files if the SEQUENCE field is
omitted in an export.

Change-Id: I313e32eb43a1032a7f33511fe1ea2b95101c0447