[BUGFIX] FileIcon RenderOption broken
We returned TYPO3's own Icon class instead of our flexible TagInterface
classes.

Unfortunately, if we want to make use of the easy internal TYPO3 API for
getting file icons, we'll have to make do with a flat string for the img
tag. At least for now, this is ok.

Change-Id: I770a40679b7aebe283c68ba80e5a0524ef90e732