[TASK] Add Breadcrumb Lock
One of our goals is a nested configuration in addition to normal
list/single configuration. However, this poses a challenge: what if a
nested configuration attempts to (re)configure and render the
breadcrumb, that you only wanted on your main template?

A lock should be able to prevent any changes to the state of the
breadcrumb. It can be set and unset publically, there's no cheating with
this, really.

The lock should -by default- also disable rendering of the VH. It can be
ignored with the ignoreLock="TRUE" attribute.

This whole setup allows us to include a crumbpath VH on all of our
templates, while still being able to only render it on the uppermost
template in case of nested-configuration, without having to resort to
different templates for this default scenario.

Change-Id: I581bd3646c366a02213e418814277691fad5435f