[TASK] Add Flexform overrides
Allow imports to be selected through a flexform. It should gain first
preference over other means of import configuration.

Also allow custom TS to be set per plugin. This overrides the entire
settings part of the configuration, so it should be complete, if
anything. Note that this is only really necessary if you ever set 2
plugins on the same page through content element plugins. (However, this
probably breaks either if the navigation remains on a single page)

Change-Id: I41344b711eb272a4db3d25168e05602cf65200f1