[BUGFIX] Other plugins within AdvancedAction break
It seems that introducing a SwitchableControllerAction in the flexform
of our DecosData plugin, we've effectively shut down our ability to run
other plugins from within our plugin, because it uses that flexform
value as a filter for allowed controllers and actions. TYPO3 obviously
doesn't expect you to run a plugin from within a plugin. Which basically
makes our ability to do so a bit of hack, but since we allow any
ContentObject type, and since there will be other ways to combine
Decosdata plugins with other plugins, I don't see the harm.

The solution for now, is to remove the SwitchableControllerAction. The
simple actions serve no purpose as of yet, they can be packaged in one
or several other plugins once I get to that.

Change-Id: I6d6b23526c6d1880680c4225f020aa326071860f