[TASK] Add note of necessity of CHash checking
As a replacement of the old SecureParameters feature, we're using a
token that is derived from the page parameters and created through the
encryptionKey. However, we already have such a token in TYPO3: CHash.
Extbase plugins can be forced to check for CHash values (default), so
we'll explicitly enable it regardless of the default value and add a
note describing its usefulness.

Still, for it to be effective, we rely on
$GLOBALS['TYPO3_CONF_VARS']['FE']['pageNotFoundOnCHashError'] = 1 being
set. Otherwise, only caching gets disabled when a page with manually
modified variables is accessed.

Change-Id: I1f4544d04fb784d21158217e8302a0023b0f3335