[TASK] Replace Default action with Multi and Single action
Default is fine as it is, but it turns out it will be much easier for
building API output if I can reliably return a single content section
without it being part of an array.

So I'm changing the way the section selection on default works: we now
have multi, and single is specifically for getting a single content
element out of multiples.

Change-Id: I6b63a57309edcc1aecb5eb9f2b1d9059bfb83fd6