[TASK] Execute content RenderOptions prior to templating
When retrieving data in apiMode, the use case is often to retrieve fully
processed data. If we only get query data, and process renderOptions
only through Fluid VHs, how do we achieve the same result when getting
data through the API? (without adding apiMode-specific overhead)

So for now, my solution is to move the RenderOptions processing (for
content) out of the templates and into the TypeProcessor. Titles can
still be processed through template.

If we ever need a mixed solution, we could apply this only on apiMode I
guess, and restore the templates to their original state. The VH will
not be removed anyway.

+ might be a bit faster than using VHs
- loses flexibility of providing tag attributes on template-provided
elements

Change-Id: I15a37f3f928988cc9028089346c98ee202c2edbe