[TASK] Add AssetProvider lib to project for impending JS needs
We're also adding a tiny "feature" that I'm not sure I want to backport
to the main lib, because our use-case for it is very decosdata-specific:

I intend to use it on a feature-base, not on a controller/action base.
So unlike with actions, it could happen that a feature gets its assets
added more than once. The change prevents double attempts to include an
asset.

Change-Id: Icbb7c1b8b49dd367c8f5f9acb3626a1284810225