[BUGFIX] Migration ValueRefs with UniqueBy are cached incorrectly
A single migration process caches ValueRefs during runtime, however,
when a UniqueBy such as 'pid' is added to the equation, the cache should
of course heed this. It doesn't. ItemType and Field records are stored
on one pid but also referenced on another if the data migrated in a
single run is on separate pages.

Change-Id: I245d97e0f7f6bf2c4f7c9b5311e04cc640324d84