[BUGFIX] Fix wrong field and type id's on import
If multiple imports on different pid's are processed in a single run,
the runtime cache of itemtypes and fields does not take into account the
difference in pid, so an itemfield for an item on pid 2 can easily
reference a field on pid 1, causing issues like double fields per item,
etc.

The fix is to include the pid in the cache keys.

Change-Id: Id0b8b5de39682a7a73b179c8544728ab9322a8fd