[TASK] Support nesting of Options
In preparation of e.g. the upcoming PdfSplit RenderOption, we'll require
the ability to apply options specifically on an array of results of a
former option.

To be able to do this, we need to safeguard index, item and content
values in the Option Service classes, making sure their values cannot be
overwritten for a different level of Options.

Change-Id: I7976e4098150b295db68abbbc1954c421657efd7