[BUGFIX] Support renderOptions on non-existing contentfields
TagBuilder no longer allows NULL values, as it should be. But this
exposes the problem that we actually do allow them through various
configs that expect it once they start applying RenderOptions. A quick
fix is to simply create the content field on the fly when a RenderOption
is applied.

Change-Id: I1e5321e50ebf94ab07b0e2d9a549399d2310a78d