[FEATURE] Add support for MockFile to RenderOptions
I don't want RenderOptions that generate potentially tons of (temp)
files on the fly to unload these on the FAL, but I want to remain using
the FAL APIs for its ease of use and standardization.

Introducing the MockFile as part of TYPO3FalApi, that extends normal
File objects but don't persist.

RenderOptions should be able to transparently and interchangably use
either File or MockFile objects through the FileHandler trait.