[TASK] Refactor CommandRunner Trait into service class
The service class needs to be able to run commands that can be set
through backend user input, without the security concerns that we'd have
by just allowing any command.

So for one, we can set allowed binaries to limit the binaries supported
by whatever class is running the command. This should allow wildcards so
that we can retain some flexibility in which commands to allow.

Another important feature would be to allow command substitutions, with
which are kind of like variables for almost complete commands, so that
we can allow potentially revealing commands like grep, and awk in only a
limited manner.