[FEATURE] Add PdfForeachPage RenderOption
As an alternative to the PdfSplit option, this allows you to run through
each page of a PDF if you only need to know the page number/index, and
not the splitted files that PdfSplit additionally provides.

This makes for a faster and less resource-hungry process in use-cases
that match the above description. (currently more than half of our
pdf-split use-cases)