[FEATURE] Add pagination for use by RenderOptions
RenderOptions should be able to paginate their results through a general
interface. I've chosen for the ability to set a 'paginate' array from
within a RenderOptions configuration element. The RenderOptionService
will detect this and call on a paginator service specific for this piece
of content, if the given option supports a a public 'paginateIterate'
method. This method will be part of the OptionInterface at some point,
when I introduce an OptionAbstract for RenderOptions.

The paginator service is going to be very general usage and should even
be able to partly replace the original paginateService (at a later
point). It can be initialized with a callback, which in the case of
options will be the aforementioned 'paginateIterate' method. This way,
options can define exactly what happens in the iterations of their
pagination logic themselves, while the paginator service takes care of
all the general pagination stuff.