[FEATURE] Support pagination in the Pdf* RenderOptions
- PdfSplit
- PdfForeachPage

.. both need to be able to withhold the immediate loading of all PDF
pages, as a way to solve the insanely bad performance of these options
when huge PDF documents are processed for the very first time.

When predecessor decospublisher was in use at its sponsors, we've seen
documents of hundreds(!!) of pages that were to be processed this way.
Without the terribly hacky preloader of decospublisher (which was
unmaintainable with every TYPO3 migration), pages would time out in any
sane webserver config. By adding pagination, we pave the way to load PDF
pages through ajax requests once we add full XHR support to the new
paginator service, as a much cleaner approach to solving the issue.