[TASK] Move content pagination configuration to contentfield level
While we configured contentfield level paging on specific renderOptions,
it's much more convenient to place it outside of the renderOption level.
It doesn't make sense to have multiple configurations on a single
contentfield anyway.

We still need a switch on the option-level to determine if the option
can call on the paginator, so we're changing the paginate array there
into a boolean.