[FEATURE] Disable XHR pagination autoload by default
Add a xhr.autoload = [0|1] configuration in paginate config blocks. This
will have to be enabled explicitly to return to the old behavior.

The new behavior will cause the xhr pager to be clickable (also as
fallback if onReach fails) at first, and only then autoload.

This also requires us to change some of the pagination logic that
expects the xhr setting to be a boolean. Since it's now an array, we've
moved the xhr boolean to:
xhr.enable = [0|1]

This is more explicit and clear, in combination with added xhr settings
starting with autoload.