[BUGFIX] Image scaling goes out of visibility for many use-cases
Generally, we scale from the top left to bottom right, but the last item
will go from bottom left to top right. If there is only 1 item, which
turns out to be often for many use-cases, it is also the last item, and
the bottom-approach causes the image to scale out of visibility.

While we can't really prevent this for all cases, we should improve the
defaults by letting it scale from the center out, so that it will work
beautifully for most use cases. Clients with special cases can then
still provide a CSS override to handle this differently.