[TASK] Add 'DB Statement = raw query' fallback
It seems the way I've been using my prepared statements, causes TYPO3 v9
to try and interpret it as a raw query, because it's not an instance of
either the TYPO3 QueryBuilder or the Doctrine DBAL Statement.

For now, we'll work around this by returning our available
getProcessedQuery() value for when it gets typecast as a string.