[TASK] Replace default field separator
By default, the default field separator is the pipe symbol, which
separates item id's in case multiple ones are used to link to a single
level. We make things a bit easier on ourselves if we change this to a
dash, which is a character which does not require any special url
encoding.

Note that this may cause a tiny compatiblity issue if realurl was used
to cache URLs with a separator in one of the request parameters.
Luckily, there has ever been only 1 implementation that used it, and
it's ditching realurl at the same time it updates this extension.