[TASK] Add advanced routing abilities
Considering the flexibility and complexity of the plugin configurations
that are possible with this ext, so too was its routing in realurl,
requiring joins and slug transformations to achieve contextually
relevant speaking paths.

These possibilities are missing from the vanilla TYPO3 v9 routing, so
the extension adds its own enhancer and aspect to provide for this.