[BUGFIX] QueryBuilder: first-added constraints cant have keys
Constraints can be added with a key, so they can easily be replaced
later on through that same key. However, the first constraint added to a
ConstraintContainer doesn't do anything with the provided key, causing
issues when attempting to replace them.

This fix ensures they can have a key so they can be identified and
replaced.